update vulnerability parsing for current master and minor refactor
Moves the vulnerability code into the sysfs package because it only uses
data from sysfs.  Also simplifies the Vulnerability struct and types.
