--- cpu/vulnerability.go
@@ -1,66 +1,0 @@
-// Copyright 2019 The Prometheus Authors
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package cpu
-
-import (
-	"fmt"
-	"strings"
-)
-
-const (
-	notAffectedSysFsValue = "Not Affected"
-	vulnerableSysFsValue  = "Vulnerable"
-	mitigationSysFsValue  = "Mitigation"
-)
-
-const (
-	NotAffected VulnerabilityState = "not affected"
-	Vulnerable  VulnerabilityState = "vulnerable"
-	Mitigation  VulnerabilityState = "mitigation"
-)
-
-type VulnerabilityState string
-
-func (s VulnerabilityState) String() string {
-	return string(s)
-}
-
-type Vulnerability struct {
-	CodeName   string
-	State      VulnerabilityState
-	Mitigation string
-}
-
-func ParseVulnerability(name, value string) (Vulnerability, error) {
-	v := Vulnerability{CodeName: name}
-
-	if value == notAffectedSysFsValue {
-		v.State = NotAffected
-		return v, nil
-	}
-
-	if strings.HasPrefix(value, vulnerableSysFsValue) {
-		v.State = Vulnerable
-		v.Mitigation = strings.TrimPrefix(strings.TrimPrefix(value, vulnerableSysFsValue), ": ")
-		return v, nil
-	}
-
-	if strings.HasPrefix(value, mitigationSysFsValue) {
-		v.State = Mitigation
-		v.Mitigation = strings.TrimPrefix(strings.TrimPrefix(value, mitigationSysFsValue), ": ")
-		return v, nil
-	}
-
-	return v, fmt.Errorf("unknown vulnerability state for %s: %s", name, value)
-}
