+++ sysfs/fs.go
@@ -14,11 +14,6 @@
 package sysfs
 
 import (
-	"io/ioutil"
-	"path/filepath"
-	"strings"
-
-	"github.com/prometheus/procfs/cpu"
 	"github.com/prometheus/procfs/internal/fs"
 )
 
@@ -45,31 +40,4 @@ func NewFS(mountPoint string) (FS, error) {
 		return FS{}, err
 	}
 	return FS{fs}, nil
-}
-
-// BcacheStats retrieves a map of vulnerability names to their mitigations.
-func (fs FS) CPUVulnerabilities() ([]cpu.Vulnerability, error) {
-	matches, err := filepath.Glob(fs.Path("devices/system/cpu/vulnerabilities/*"))
-	if err != nil {
-		return nil, err
-	}
-
-	vulnerabilities := make([]cpu.Vulnerability, 0, len(matches))
-	for _, match := range matches {
-		name := filepath.Base(match)
-
-		value, err := ioutil.ReadFile(match)
-		if err != nil {
-			return nil, err
-		}
-
-		v, err := cpu.ParseVulnerability(name, strings.TrimSpace(string(value)))
-		if err != nil {
-			return nil, err
-		}
-
-		vulnerabilities = append(vulnerabilities, v)
-	}
-
-	return vulnerabilities, nil
 }
